% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstut.R
\name{ions}
\alias{ions}
\alias{mstut}
\alias{analyse}
\alias{detect}
\alias{spectrum}
\title{Creates an object of class \code{ions}}
\usage{
ions(npeaks = 10, mzrange = c(100, 1000), nimg = 100)

analyse(x, sleep = 0.1)

detect(x, new = FALSE)

spectrum(x, ...)
}
\arguments{
\item{npeaks}{A \code{numeric} scalar defining the number of unique
peaks (M/Z values). Default is 10.}

\item{mzrange}{A \code{numeric} of length 2 defining the range of
possible M/Z values. Default is \code{c(100, 1000)}.}

\item{nimg}{A \code{numeric} scalar. When analysing the ions, their
separation along their M/Z values will be split along a
sequence of length \code{nimg}. Default is 100.}

\item{x}{An object of class \code{ions}.}

\item{sleep}{How much time to wait before producing the next plot.}

\item{new}{A \code{logical} scalar, indicating if the separated ions
(last frame of calling \code{analyse) should be plotting, or whether the detection should be overlaid. Default is}FALSE`,
to add the plot on top of the opened device.}

\item{...}{Additional arguments passed to \code{\link[graphics:plot]{graphics::plot()}}.}
}
\value{
An object of class \code{ions}.

Use for it's side effect.
}
\description{
This is the constructor function to generate a set of ions that
can later be analysed with \code{analyse()} and detected with
\code{detect()}.
}
\examples{
set.seed(1L)
x <- ions(nimg = 5)
x
analyse(x)
detect(x)
spectrum(x)
}
\author{
Laurent Gatto
}
